refactor: break cli client import cycle
